test(publisher): split receipt-hint suite into focused files behind a shared harness
r21 (3878490032): async-lift-receipt-hint.test.ts crossed 1,000 lines. The parked
one-wallet and two-wallet scenario builders, polling helpers, and the wallet-lock
assertion move to test/_helpers/receipt-hint-scenario.ts behind one factory; the
23 rows split into core/evidence, scheduling/fairness, and failure/retry suites,
every row body preserved verbatim (assertion-level identity verified against the
pre-split file). The new helper joins the typechecked helper lane, which surfaced
three latent stub-typing gaps in the relocated builders; fixed without changing
any runtime observable (operationKind is unread by recordRpcAccepted; the execute
stub keeps its historical undefined settle value, documented at the cast).
